+++ appListarClientes/src/bootstrap.js
@@ -1,0 +1,15 @@
+import React from "react";
+import ReactDom from 'react-dom/client';
+import { BrowserRouter } from "react-router-dom";
+import {App} from "./components/App";
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+const root = ReactDom.createRoot(
+    document.getElementById('root'),
+);
+
+root.render(
+    <BrowserRouter>
+        <App />
+    </BrowserRouter>
+);
